Cambios

+++ PTS/src/main/java/pe/edu/upc/pts/dtos/RolDTO.java
@@ -1,14 +1,14 @@
 package pe.edu.upc.pts.dtos;
 
 public class RolDTO {
-    private int idRol;
+    private Long idRol;
     private String nameRol;
 
-    public int getIdRol() {
+    public Long getIdRol() {
         return idRol;
     }
 
-    public void setIdRol(int idRol) {
+    public void setIdRol(Long idRol) {
         this.idRol = idRol;
     }
